requirements.txt
requests
pdfplumber
python-dotenv 
google-genai
google-generativeai
supabase
openai
groq